image slider fix
