+++ src/features/public/Donations/components/SingleProjectDetails.tsx
@@ -95,28 +95,6 @@ function SingleProjectDetails({
     setOpen(true);
   };
 
-  let projectImages: { content: () => JSX.Element }[] = [];
-
-  React.useEffect(() => {
-    project.images.forEach((image: any) => {
-      let imageURL = loadImageSource(image.image);
-      projectImages.push({
-        content: () => (
-          <div
-            className={styles.projectImageSliderContent}
-            style={{
-              background: `linear-gradient(to top, rgba(0,0,0,1), rgba(0,0,0,0.2), rgba(0,0,0,0), rgba(0,0,0,0)),url(${imageURL})`,
-            }}
-          >
-            <p className={styles.projectImageSliderContentText}>
-              {image.description}
-            </p>
-          </div>
-        ),
-      });
-    });
-  }, [project]);
-
   const ProjectProps = {
     project: project,
   };
@@ -267,7 +245,7 @@ function SingleProjectDetails({
               ) : null}
               <div className={styles.projectImageSliderContainer}>
                 {project.images.length > 0 ? (
-                  <ImageSlider project={projectImages} />
+                  <ImageSlider project={project} />
                 ) : null}
               </div>
               {/* {infoProperties ? <ProjectInfo infoProperties={infoProperties} /> : null}
